requirements.txt
streamlit==1.52.1
numpy==2.3.5
Pillow==12.0.0
tensorflow==2.20.0
h5py==3.15.1
